Use tmp_path in ghostwriter tests

+++ hypothesis-python/scripts/other-tests.sh
@@ -31,7 +31,9 @@ $PYTEST tests/redis/
 pip uninstall -y redis fakeredis
 
 $PYTEST tests/typing_extensions/
-pip uninstall -y typing_extensions
+if [[ "$HYPOTHESIS_PROFILE" != "crosshair" ]]; then
+  pip uninstall -y typing_extensions
+fi
 
 if [ "$(python -c 'import sys; print(sys.version_info[:2] >= (3, 9))')" = "True" ] ; then
   pip install "$(grep 'annotated-types==' ../requirements/coverage.txt)"
@@ -62,14 +64,10 @@ if [ "$(python -c $'import platform, sys; print(sys.version_info.releaselevel ==
     pip install "$(grep 'numpy==' ../requirements/coverage.txt)"
   fi
 
-  case "$(python -c 'import platform; print(platform.python_implementation())')" in
-    PyPy|GraalVM)
-      ;;
-    *)
-      $PYTEST tests/array_api
-      $PYTEST tests/numpy
-  esac
-
   $PYTEST tests/ghostwriter/
-  pip uninstall -y black numpy
+  pip uninstall -y black
+
+  if [ "$(python -c "import platform; print(platform.python_implementation() not in {'PyPy', 'GraalVM'})")" = "True" ] ; then
+    $PYTEST tests/array_api tests/numpy
+  fi
 fi
